pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-data-team"
version = "0.1.0"
description = "Email-driven autonomous data science swarm for dataset attachments and archive scoring."
requires-python = ">=3.10"
dependencies = [
  "agentmail>=0.4.10",
  "beautifulsoup4>=4.12.0",
  "crewai>=1.11.0",
  "crewai-tools>=1.11.0",
  "httpx>=0.28.0",
  "markdown-it-py>=3.0.0",
  "openai>=1.0.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "python-dotenv>=1.0.0",
  "scikit-learn>=1.4.0",
  "tavily-python>=0.5.0",
  "tenacity>=8.2.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
autonomous-data-team = "autonomous_data_team.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["autonomous_data_team*"]
